% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oasi_fetch_data.R
\name{fetch_locations_data}
\alias{fetch_locations_data}
\title{Fetch Locations Data}
\usage{
fetch_locations_data(domain)
}
\arguments{
\item{domain}{A character string representing the domain for which to fetch location data.}
}
\value{
A data frame containing the locations data if the request is successful; otherwise, NULL.
}
\description{
This function retrieves location data from the OASI API based on the specified domain.
}
\examples{
\dontrun{
locations <- fetch_locations_data("example_domain")
print(locations)
}
}
