% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oasi_fetch_data.R
\name{fetch_parameters_data}
\alias{fetch_parameters_data}
\title{Fetch Available Parameters for a Location}
\usage{
fetch_parameters_data(domain, location_code)
}
\arguments{
\item{domain}{A character string representing the domain for which to fetch parameters.}

\item{location_code}{A character string representing the location code for which to fetch parameters.}
}
\value{
A data frame containing the available parameters if the request is successful; otherwise, NULL.
}
\description{
This function retrieves the available parameters for a specified location from the OASI API.
}
\examples{
\dontrun{
parameters <- fetch_parameters_data("example_domain", "location_code")
print(parameters)
}
}
