% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oasi_fetch_data.R
\name{fetch_time_series_data}
\alias{fetch_time_series_data}
\title{Fetch Time Series Data}
\usage{
fetch_time_series_data(
  domain,
  location_code,
  parameter,
  resolution,
  from_date,
  to_date
)
}
\arguments{
\item{domain}{A character string representing the domain for which to fetch data.}

\item{location_code}{A character string representing the location code for which to fetch data.}

\item{parameter}{A character string representing the parameter to fetch.}

\item{resolution}{A character string representing the resolution of the data (e.g., "daily", "hourly").}

\item{from_date}{A character string representing the start date for the data (in "YYYY-MM-DD" format).}

\item{to_date}{A character string representing the end date for the data (in "YYYY-MM-DD" format).}
}
\value{
A character string containing the time series data in CSV format if the request is successful; otherwise, NULL.
}
\description{
This function retrieves time series data for a specified parameter, location, and time range from the OASI API.
}
\examples{
\dontrun{
data <- fetch_time_series_data("example_domain", "location_code", "parameter", "daily", "2024-01-01", "2024-01-31")
print(data)
}
}
