% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputation.R
\name{impute_missing_data_with_model}
\alias{impute_missing_data_with_model}
\title{Impute Missing Data Using Trained Models}
\usage{
impute_missing_data_with_model(data_frame, models)
}
\arguments{
\item{data_frame}{A data frame containing the data with missing values (\code{NA}).}

\item{models}{A list of pre-trained models corresponding to each column in the data frame.}
}
\value{
A data frame with missing values imputed.
}
\description{
This function imputes missing values in a data frame using pre-trained models.
}
\examples{
\dontrun{
imputed_data <- impute_missing_data_with_model(my_data, models)
}
}
