% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/training.R
\name{train_models}
\alias{train_models}
\title{Train Predictive Models}
\usage{
train_models(data_frame, method = "glm", glm_family = gaussian())
}
\arguments{
\item{data_frame}{A data frame containing the data to train the models on.}

\item{method}{A character string specifying the model type. Options are \code{"lm"} for linear models and \code{"glm"} for generalized linear models. Defaults to \code{"glm"}.}

\item{glm_family}{A family function to be used if \code{method} is set to \code{"glm"}. Defaults to \code{gaussian()}.}
}
\value{
A list of trained models, one for each column in the data frame.
}
\description{
This function trains predictive models based on the provided data frame. It can use different methods such as \code{lm} or \code{glm}
to fit models for each column in the data frame. The trained models can then be used for imputation of missing data.
}
\examples{
\dontrun{
models <- train_models(my_data, method = "glm", glm_family = gaussian())
}
}
